% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{create_project}
\alias{create_project}
\title{Reserved for future use}
\usage{
create_project(name, id, description = NULL, clientRequestToken = NULL,
  tags = NULL)
}
\arguments{
\item{name}{Reserved for future use.}

\item{id}{Reserved for future use.}

\item{description}{Reserved for future use.}

\item{clientRequestToken}{Reserved for future use.}

\item{tags}{The tags created for the project.}
}
\description{
Reserved for future use. To create a project, use the AWS CodeStar console.
}
\examples{

}
