% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{describe_project}
\alias{describe_project}
\title{Describes a project and its resources}
\usage{
describe_project(id)
}
\arguments{
\item{id}{[required] The ID of the project.}
}
\description{
Describes a project and its resources.
}
\section{Accepted Parameters}{
\preformatted{describe_project(
  id = "string"
)
}
}

