% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{describe_user_profile}
\alias{describe_user_profile}
\title{Describes a user in AWS CodeStar and the user attributes across all projects}
\usage{
describe_user_profile(userArn)
}
\arguments{
\item{userArn}{[required] The Amazon Resource Name (ARN) of the user.}
}
\description{
Describes a user in AWS CodeStar and the user attributes across all projects.
}
\section{Accepted Parameters}{
\preformatted{describe_user_profile(
  userArn = "string"
)
}
}

