% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{list_projects}
\alias{list_projects}
\title{Lists all projects in AWS CodeStar associated with your AWS account}
\usage{
list_projects(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The continuation token to be used to return the next set of results, if the results cannot be returned in one response.}

\item{maxResults}{The maximum amount of data that can be contained in a single set of results.}
}
\description{
Lists all projects in AWS CodeStar associated with your AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_projects(
  nextToken = "string",
  maxResults = 123
)
}
}

