% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{list_resources}
\alias{list_resources}
\title{Lists resources associated with a project in AWS CodeStar}
\usage{
list_resources(projectId, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{projectId}{The ID of the project.}

\item{nextToken}{The continuation token for the next set of results, if the results cannot be returned in one response.}

\item{maxResults}{The maximum amount of data that can be contained in a single set of results.}
}
\description{
Lists resources associated with a project in AWS CodeStar.
}
\examples{

}
