% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{list_team_members}
\alias{list_team_members}
\title{Lists all team members associated with a project}
\usage{
list_team_members(
  projectId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{projectId}{[required] The ID of the project for which you want to list team members.}

\item{nextToken}{The continuation token for the next set of results, if the results cannot be returned in one response.}

\item{maxResults}{The maximum number of team members you want returned in a response.}
}
\description{
Lists all team members associated with a project.
}
\examples{

}
