% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{list_user_profiles}
\alias{list_user_profiles}
\title{Lists all the user profiles configured for your AWS account in AWS CodeStar}
\usage{
list_user_profiles(
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{nextToken}{The continuation token for the next set of results, if the results cannot be returned in one response.}

\item{maxResults}{The maximum number of results to return in a response.}
}
\description{
Lists all the user profiles configured for your AWS account in AWS CodeStar.
}
\examples{

}
