% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_package.R
\docType{package}
\name{paws.codestar-package}
\alias{paws.codestar}
\alias{paws.codestar-package}
\title{paws.codestar: AWS CodeStar}
\description{
This is the API reference for AWS CodeStar. This
    reference provides descriptions of the operations and data types for
    the AWS CodeStar API along with usage examples.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
