% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{tag_project}
\alias{tag_project}
\title{Adds tags to a project}
\usage{
tag_project(id, tags)
}
\arguments{
\item{id}{The ID of the project you want to add a tag to.}

\item{tags}{The tags you want to add to the project.}
}
\description{
Adds tags to a project.
}
\examples{

}
