% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{untag_project}
\alias{untag_project}
\title{Removes tags from a project}
\usage{
untag_project(id, tags)
}
\arguments{
\item{id}{[required] The ID of the project to remove tags from.}

\item{tags}{[required] The tags to remove from the project.}
}
\description{
Removes tags from a project.
}
\section{Accepted Parameters}{
\preformatted{untag_project(
  id = "string",
  tags = list(
    "string"
  )
)
}
}

