% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{update_project}
\alias{update_project}
\title{Updates a project in AWS CodeStar}
\usage{
update_project(
  id = "string",
  name = "string",
  description = "string"
)
}
\arguments{
\item{id}{[required] The ID of the project you want to update.}

\item{name}{The name of the project you want to update.}

\item{description}{The description of the project, if any.}
}
\description{
Updates a project in AWS CodeStar.
}
\examples{

}
