% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codestar_operations.R
\name{list_tags_for_project}
\alias{list_tags_for_project}
\title{Gets the tags for a project}
\usage{
list_tags_for_project(
  id = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{id}{[required] The ID of the project to get tags for.}

\item{nextToken}{Reserved for future use.}

\item{maxResults}{Reserved for future use.}
}
\description{
Gets the tags for a project.
}
\examples{

}
